Module user::syscall

All syscalls of core-os

Syscalls of core-os are defined and implemented with Rust primitives (e.g. str, [u8]). This module will transmute these Rust primitives into pointers and other machine-specific representations before calling functions in syscall_internal and trapping into kernel.

Usage of syscalls is listed in their corresponding sub-page.

Constants

EXEC_MAX_ARGS

Functions

close
Close a file with file descriptor fd.

dup
Duplicate file descriptor fd.

exec
Replace current process image with the new one in the filesystem.

exit
Exit current process with exit code code.

fork
Fork current process.

open
Open file of path with mode.

read
Read content from file descriptor fd.

wait

write
Write content to file descriptor fd.