Module user::syscall_internal

Internal representation of syscall

As Rust primitives can't be directly transferred between kernel space and user space, all Rust primitives will be transmuted into pointers in syscall module, and then this module will finally trap into kernel.

Functions

__close⚠

__dup⚠

__exec⚠

__exit⚠

__fork⚠

__open⚠

__read⚠

__wait⚠

__write⚠