Crate user

Helper utilities for switching modes in SetUID applications.

Macros

as_effective

as_real

run_as

Structs

Error

An error when trying to change UID/GID.

Sync

A synchronization primitive. Only one thread will ever have a Sync object. When the Sync object drops, control is relinquished to another thread.

Enums

Mode

A SetUID mode.

Statics

GROUP

The Real, Effective, and Saved GID of the application.

SETUID

Whether the system is actually running under SetUid. If false, all functions here are no-ops.

USER

The Real, Effective, and Saved UID of the application.

Functions

current

Get the current user mode

drop

Destructively change mode, preventing the process from returning. This function will set Real, Effective, and Saved values to the desired Mode. This prevents the process from changing their mode

obtain_lock

restore

Restore a saved Uid/Gid combination This function fails if the syscall does.

revert

Revert the Mode to the original. This function returns to the values of USER and GROUP. This function can fail if the underlying syscall does.

set

Set the Mode. This function can never be misused to lock out the process from returning to the original, or any other mode. This function returns an error if the mode could not be changed. Otherwise, it returns the old mode for use in restore()