user

Function set

pub fn set(mode: Mode) -> Result<(ResUid, ResGid), Errno>

Set the Mode. This function can never be misused to lock out the process from returning to the original, or any other mode. This function returns an error if the mode could not be changed. Otherwise, it returns the old mode for use in restore()

Examples

user::set(user::Mode::Real).unwrap();

Notes

user::set(Mode::Original) is functionally identical to user::revert()

user::set(Mode::Existing) is a no-op.